Module wasmer_interface_types::decoders::binary

Parse the WIT binary representation into an AST.

Functions

adapters
Parse a list of adapters.

byte
Parse a byte.

exports
Parse a list of exports.

implementations
Parse a list of implementations.

imports
Parse a list of imports.

instruction
Parse an instruction with its arguments.

interfaces
Parse complete interfaces.

list
Parse a list, with an item parser.

parse
Parse a sequence of bytes, expecting it to be a valid WIT binary representation, into an Interfaces structure.

record_type
Parse a record type.

string
Parse a UTF-8 string.

ty
Parse an interface type.

types
Parse a list of types.

uleb
Parse an unsigned Little Endian Based (LEB) with value no larger than a 64-bits number. Read LEB128 to learn more, or the Variable Length Data Section from the DWARF 4 standard.