Module wasmer_interface_types::decoders::wat

Parse the WIT textual representation into an AST.

Modules

keyword

Structs

AtInterface

Buffer
A low-level buffer of tokens which represents a completely lexed file.

Enums

FunctionType

Interface

Functions

parse
Parse a WIT definition in its textual format, and produces an AST with the Interfaces structure upon succesful.