Module wasmer_interface_types::interpreter

A stack-based interpreter to execute instructions of WIT adapters.

Modules

instruction

instructions

stack
A very light and generic stack implementation, exposing only the operations required by the interpreter.

wasm
An hypothetic WebAssembly runtime, represented as a set of enums, types, and traits —basically this is the part a runtime should take a look to use the wasmer-interface-types crate—.

Structs

Interpreter
An interpreter is the central piece of this crate. It is a set of executable instructions. Each instruction takes the runtime as argument. The runtime holds the invocation inputs, the stack, and the WebAssembly instance.

Runtime
Represents the Runtime, which is used by an adapter to execute its instructions.

Enums

Instruction
Represents all the possible WIT instructions.

Type Definitions

ExecutableInstruction
Type alias for an executable instruction. It's an implementation details, but an instruction is a boxed closure instance.