Function wasmer_interface_types::interpreter::instructions::records::record_lift_

fn record_lift_(
    stack: &mut Stack<InterfaceValue>, 
    record_type: &RecordType
) -> Result<InterfaceValue, InstructionErrorKind>

Build an InterfaceValue::Record based on values on the stack.

To fill a record, every field field_1 to field_n must get its value from the stack with value_1 to value_n. It is not possible to use Stack::pop because the one-pass algorithm does not know exactly the number of values to read from the stack ahead-of-time, so Stack::pop1 is used. It implies that values are read one after the other from the stack, in a natural reverse order, from value_n to value_1. Thus, the values vector must be filled from the end to the beginning. It is not safely possible to fill the values vector with empty values though (so that it is possible to access to last positions). So a VecDeque type is used: it is a double-ended queue.