Module wasmer_interface_types::interpreter::stack

A very light and generic stack implementation, exposing only the operations required by the interpreter.

Structs

Stack
A stack implementation of the Stackable trait, based on a vector.

Traits

Stackable
The Stackable trait represents a small basic set of operations required by the interpreter.